Fix computation of overridden methods.
Each JMethod contains a list of the methods it overrides; such
list should be exhaustive but the previous computation would
not include all overriden methods.

In the following example

  public class A {
    public void m() { }
  }

  public interface I {
    void m();
  }

  public class B extends A implement I {
  }

A.m() overrides I.m() but was not recorded in the JMethod
representation of A.m().

This information is mostly used in optimizations resulting in
at best missed optimization opportunities and subtle bugs at
worst.

This patch also makes the list of overriden methods in JMethod
updated when prunning.

Change-Id: I1f9af8df529f48b021c8c5de1290780ab5c0f645