Update patch set 2
Patch Set 2:

The new line emaits by the converter at the beginning of the file is fixed: https://gwt-review.googlesource.com/#/c/11140/

Concerning the comments, unfortunately, Flute skips the comments during the parsing. There is a method comment() in the DocumentHandler interface (kind of callback call by flute during the parsing) but this method is never called by Flute. Flute is open source so we could maybe modify it but I don't know if it' worth trying

Patch-set: 2
Label: Verified=0