Update patch set 1
Patch Set 1:

(1 comment)

they're not actually compilable as libraries now because they still contain dependency problems. what's actually changing is that we no longer support library compiles (only monolithic compiles) and as a result there's no reason to distinguish between modules that are and are not correct enough to be individually compiled.

Patch-set: 1
Label: Verified=0