Fixes bug in (incremental) link time pruning of bridge methods.
Link time pruning in incremental compiles relies on a combination of 
recorded control flow indexes and RTA execution. The RTA execution 
makes the simplifying assumption that any method will know its complete
transitive set of overiding methods and the old control flow recording 
reduced granularity and saved some space by indexing methods by their 
signature without return type. This was resulting in grouping together 
of a method and it's samely named bridge method and was believed safe.

But actually this grouping was invalidating the assumption that a 
method would know its complete transitive set of overiding methods, 
since this method signature grouping was effectively adding more 
overiders without also updating the overridden method's indexes. The
net effect of this violation was to occasionally fail to see that some 
method is reachable and that could cascade into failing to see that
some type was instantiated, thus output could miss a whole type and
result in a runtime error.

Stopping the grouping by increasing the granularity of recorded method 
signatures to include return type fixes the bug and will also slightly 
improve the accuracy of the pruning and thus slightly decrease output 
size. Though this will come at the cost of a little bit larger size for
saved MinimalRebuildCache instances.

Change-Id: I7a0c6bb12bae69f6eb8680002721a81caa3f8ea0