Fix unnecessary scan for absent classes.
During compilation class references are resolved by JDT using
compiled bytecode. If the class is not part of the current compilation,
JDT will query a NameEnvironment for the bytocode of the class which
GWT might provide from its caches. In the case that is not in any cache
GWT falls back to loading from the classpath (which is actually only
allowed for annotations). In such cases the result is not cached forcing
resource scans every time (even if the class is not found at all).

This patch solves the issue by caching the results of such queries.

Change-Id: I92e78da8e39fb4f4a2321cac45d43ee6e69467b4