Enhances I18NCreator for creating Message interfaces from properties files
Currently I18NCreator creates method signatures that only use String
arguments regardless of any MessageFormat that may be present for
an argument. This patch attempts to create more accurate interface
signatures based on the MessageFormat and also includes some GWT specific
formats such as {0,list},{0,localdatetime,..} and Plurals.

This patch also proposes to add the convention {0,safehtml} to mark that
an argument to be of type SafeHtml.  If an Argument is found to be of
type SafeHtml the return type of the method will be set to SafeHtml.

Issue 9114 was created in the tracker to cover this patch

Issue 8943 is also fixed with this patch by adding the @param annotations
for each argument

Change-Id: I9f8217ab6ca211a1ebba2ff2010df463e3a682cf