Update patch set 2
Patch Set 2:

I don't totally agree, especially because instanceof behavior is defined by Javascript; the platform has made the choice of scoping it by frame, not us. However I don't like any potential surprise with it either; these are fair points so I put them to the doc as a comment to the related part and we can discuss that more again.

My main rational for starting with the simpler choice is to not lock ourselves to the decision;  when instanceof starts working across frames in a never GWT version, even it is a breaking change, in practice shouldn't cause issues for most people.

Patch-set: 2