Update patch set 2
Patch Set 2:

> One comment I've gotten BTW today is that j2objc manages to preserve Java semantics but GWT doesn't and this has caused headaches.

To be fair, underlying platform for ObjectiveC is a saner platform and closer to java and they don't try to do crazy optimizations as we need to do :)

I wonder which particular semantic differences caused headaches? They are not referring to missing stuff like reflection, right?

Patch-set: 2
Label: Verified=0