Always use basetype for @Shared replacements.
We were using the a possible subtype of the CssResource
interface for calculating CssResource naming for shared methods.
In an incremental compilation that type does not need to be available.
Thus we were accidentally using a different name for styles making
selectors not match in recompiles that only touched one subtype.

By using the base CssResource that defined the shared method as the name
we make sure that this name will always match.

Change-Id: I4c758a42bc5543505d2cea4f18d98f318a154068