Fix missing clickevent when button was disabled on mouseover.
When the mouse cursor moved over the button while disabled a click on
the re-enabled button was ignored. This led to a situation where the
user had to move the cursor off the button and back over the button
to be able to click the button.

The implementation allowed only a click when the the hovering face was
previously applied to the button. But actually it is better to check if
the eventTarget is the button itself or a child element.

Bug: issue 8684
Change-Id: I4a2b2215bf80663dd2b12148c75d15ca91618411