Fixes swallowing of click event when CustomButton was disabled on mouseover event.
When the mouse cursor moved over the button while disabled a click on the re-enabled button was ignored. This led to a situation where the user had to move the cursor off the button and back over the button to be able to click the button.

The implementation allowed only a click when the the hovering face was previously applied to the button. But actually the click event only occurs in combination with isCapturing=true when the mouse cursor points on the button. So the condition isHovering() is not required.

Change-Id: I4a2b2215bf80663dd2b12148c75d15ca91618411