Update patch set 2
Patch Set 2: Code-Review+1

> I think the last .icobin file can be removed.

Well, it's what led the samples to have a favicon.ico too. If we remove it (which is probably a good idea given our GWT Logo Use Policy), we should maybe also remove the favicon.ico from the samples (just in case someone uses one as the basis for his project; I think we've seen spin-offs of the showcase already).

Patch-set: 2
Label: Code-Review=+1