Fix and centralize override computations.
Each JMethod contains a list of the methods it overrides; such
list should be exhaustive but the previous computation would
not include all overriden methods.

In the following example

  public class A {
    public void m() { }
  }

  public interface I {
    void m();
  }

  public class B extends A implement I {
  }

A.m() overrides I.m() but was not recorded in the JMethod
representation of A.m() (we will call this type of override
accidental).

This information is mostly used in optimizations resulting in
at best missed optimization opportunities and subtle bugs at
worst.

This patch:
  1) fixes the override computation in UnifyAST to take into
     account package-private methods and accidental override.
  2) centralized override information in JMethod. Now JMethod
     holds both overriden methods and overriding methods.
  3) updates the list of overrides in JMethod when pruning.

Change-Id: I1f9af8df529f48b021c8c5de1290780ab5c0f645