Update patch set 1
Patch Set 1:

Currently, there's an early-out in Devirtualizer

 if (instanceType instanceof JInterfaceType && ((JInterfaceType) instanceType).isJsType()) {
      return;
 }

This seems to suggest if the method call isn't tightened to Object.hashCode() but left as Foo.hashCode(), it won't get devirtualized.

Also, won't Java8 default interface methods require a form of devirtualization? If A @JsType interface has a default method, and it's invoked, the will insert a trampoline to invoke a static method if the prototype doesn't have it.

Patch-set: 1