Fixes bugs and adds tests for @JsProperty naming.
There was a bug where users were being allowed to define getters (that 
started with "get") that returned void and there was another bug where 
getters were allowed to be named like setters and setters were allowed 
to be named like getters.

I corrected these bugs and added tests to show that each of the various
disallowed forms now result in a compile error.

Change-Id: I800e3842b23ad3c4becbe5dfa1d89131ec00918a