Fixes 3 bugs in JsProperty and adds tests.
Getters of the form "getFoo()" were being improperly allowed to be 
declared to return void. Fixed.

Is functions of the form "isFoo()" were being improperly translated to 
JS as reads and writes of the "oo" property instead of the "foo" 
property, which lead to failures to round trip values between 
"setFoo()" and "isFoo()". Fixed.

Has function translation to JS was broken since "bar.hasFoo()" became
"foo in bar" rather than "'foo' in bar". This slight syntax error 
lead to undefined property errors at runtime for any property that did
not yet have a value.

Change-Id: I24cedfd4ff1beb085cd8343a5a8966e34d3c6b02