Removes incorrect JStype check from isJso.
In UnifyAst, the isJso check also tries to include
JsType but only looks at immediate interfaces, not
concerete JsType or its super classes. But fixing
that caused issues, which let me investigating its
uses.

It is used in three places:
1. Checking rebind results are not JSOs: this use is
actually incorrect because there is nothing wrong
in JsType being a rebind result. Added a test case to
reproduce the problem.

2. assimulateSourceUnits: This already explicitly
checks for JsTypes and mark them as instantiated.

3. instantiate: This use is to prevent shortcuting
types that are only reference only and doesn't
require devirtualization; JsType doesn't require
devirtualization so check doesn't apply to JsType
in this context.

So I decided to remove this code instead.

Change-Id: I2ea2e7bd47dbe100cb2866e825a7eca720ea9ea8
Review-Link: https://gwt-review.googlesource.com/#/c/11650/