Fixes incorrect obfuscation of names for JsInterop.
As old browsers doesn't support keyword as property names
GWT may mangle the name even if it is marked as unobfuscatable.

The only browser that doesn't support this is IE8[1] and
IE8 is not supported by JsInterop. For that reason, this
patch adds ability to enforce names even they are known keywords.

The patch is little bit hacky but it is a surgical change and I put
some assertions to prevent misuses.

Alternative could be using braced property access (obj["..."])
whenever the property/function name is a keyword.
That would at least keep IE8 successfully parse the script though
that looks like a much bigger change.

[1] http://kangax.github.io/compat-table/es5/#Reserved_words_as_property_names

Change-Id: If55965b09f2f7892cabf88830c1fa91f59ef4f22