Turn GSS on by default.
In GWT 2.8, GSS will be the default way to manage the CSS of a
GWT application.
We will still support the old syntax by using the automatic
conversion (will be decomissionned in GWT 3.0)
This change set GssResourceGenerator as default generator for
CssResource and remove the unused CssResourceGenerator.
It cleans also all classes that were used by the old Generator
and cleans the related tests.

Change-Id: I3495f40c4f9ab3803296586f6c88759f01257417