Cleans up JsInterop related code in UnifyAst
This also fixes the problem where we don't always correctly
collect the root types (i.e. in case of non-class-wide JsExport).

This is not a complete cleanup as there are still unnecessary
checks in UnifyAst. I want to reduce the likelihood of behavior
change in this patch.

Change-Id: Ia8e9594e028df7977fd83e4e403dbbc8bcf63a35
Review-Link: https://gwt-review.googlesource.com/#/c/11651/