Update patch set 1
Patch Set 1:

Quoted properties won't work with Closure Integration since you have to be consistent in how you access (either ALWAYS dotted, or ALWAYS quoted)

I'm ok with the solution here, but a slight nitpick on the implementation. I suppose Roberto won't object, but this tends to bother me:

We introduce a new boolean field to JsNameRef, one of the JS AST nodes with the highest occurrence. Due to memory layout alignment, packing, and 64-bit VM, this could add an extra 8 bytes to every instance, which could add a few MB to the heap usage.

An alternate, but more complex implementation would have JsProperty methods "whitelist" the nodes with bad names in a global Set instead, so you'd have O(# property methods) * set overhead vs O(# of JsNameRefs) * boolean field overhead.

It might turn out to be not that big of a win, but there you go.

Patch-set: 1