Update patch set 1
Patch Set 1:

> > That's not always possible because you don't always control the
 > > interfaces, especially in a multiplatform shared codebase. There
 > > needs to be a way to blacklist methods.
 > 
 > You need to have some control to make the interface a JsType,
 > right? If it is multiplatform codebase, perhaps you should go with
 > GwtIncompatible to remove those parts out.
 > 
 > I would like to avoid adding another skip concept as the general
 > model continuously is getting complicated to grasp (and also
 > corresponding implementations as well). So I think it should be
 > really the last resort.

We just got hit with this last Friday where @JsNoExport allowed a 27% reduction in code size. Basically, there's an interface with a method, shared by 4 platforms as well as a webworker, that returns an class. This code is called by Android, Server, iOS, and the Web Worker, but not from the main JS. However, the @JsType rescued the method and therefore pulled in massive amounts of code that for some reason, could not be pruned by the Closure Compiler.

Yes, you could do a whole bunch of refactoring to avoid this, but this means GWT is imposing a severe cost on code layout for other platforms like Android and iOS that do not need it.

@GwtIncompatible won't work because it's still needed by the WebWorker.

I think being able to blacklist methods out of a JsType is going to be needed. GWT Exporter evolved over 7 years to support this, because hundreds of apps using it found they needed stuff like that (and ExportOverlay).  When we add the ability to have a Proguard like config file it makes it even more conspicuous if you can't do it.

Patch-set: 1
Reviewer: Gerrit User 5001 <5001@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0