Update patch set 1
Patch Set 1: Code-Review+2

By always, I mean on a per property basis. So for example, if in one JS module, it is referred to as "x.import" and in another it is referred to as "x['import']", closure will obfuscate and rename x.import, but it will not do it do x['import']. So basically, all references to 'import' must be either dotted, or quoted, but not both.

Patch-set: 1
Reviewer: Gerrit User 5001 <5001@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+2