Update patch set 1
Patch Set 1:

Yes you are correct.  What I meant to say was, this patch reverts an earlier patch https://gwt-review.googlesource.com/10520 .  I want this earlier patch reinstated.  I will post that on the 10520 page. I assume there is an issue for this in the GWT bug/issue tracker, but I cannot find it.

Patch-set: 1