Update patch set 1
Patch Set 1:

In order for code to be accepted in GWT we need to be sure that there are not copyright issues with the code. For JDK specific things you should look at the Javadoc only and implement the same contract that is specified there. You should NEVER copy any code and change it.
For legal reasons it has to be a from scratch implementation.

Patch-set: 1