Fix a bug in Lambda with block body and captured outer variables.
When a lambda expression has a block method body, the variable
binding for captured outer variables are not resolved correctly.

For example, {int x = 0; () -> {int t = x; return t;}; },
the variable binding to x inside the lambda body is not resolved
correctly.

The reason is that when we resolved the variable binding in lambda
expression, we check if scope.isLambdaScope(). However, only
MethodScope may return true, and BlockScope (here
{int t = x; return t;}) always returns false.

Change-Id: I974d7fa0e40bff08493195c7293b76d2668dcbb8