Fix bugs in Lambda with nested scopes.
There are two places where we tried to find outer variable bindings
of a lambda in current method context but we did it wrong.

1). In GwtAstBuilder.resolveNameReference,
When a lambda expression has a block method body,
For example, {int x = 0; () -> {int t = x; return t;}; },
a wrong path was executed. The reason is that for the condition
scope.isLambdaScope(), only MethodScope may return true,
and BlockScope (here {int t = x; return t;}) always returns false.

2). In GwtAstBuilder.replaceLambdaWithInnerClassAllocation,
For nested lambdas, for example,
x = 0;
call(sam1 -> {call(sam2 -> {x;});});
LocalVariableBinding of x in the inner lambda (sam2->{x;})
is different from the binding of x in the outer lambda.
So we cannot get the JVariable from the outer lambda. But they
refer to the same actual outer variable. So to fix it, we compare
the actual outer variable to get the corresponding variable.

Bug: issue 9038.
Change-Id: I974d7fa0e40bff08493195c7293b76d2668dcbb8