Merges XYZpart suites and divide RequestFactory suite.
We no longer need manual grouping of the tests as we enabled
sharding.
Sharding breaks RequestFactory suite due to some tests
in EditorTest was interferring with each other in an untrivial
way so I just created a seperate Suite for RequestFactory/Editor
integration.

Also this patch enables the sharding for dev/CompilerTest.

Change-Id: I628c60d9b1d95e1faa4ab7add1ba4204c43ae1a0