Enhances TypeTightener with ability to lock into exact types.
Being able to lock to an exact type helps us to mark even more
methods as non-polymorphic which provides new optimization
opportunities.

Earlier, for example, if there was any reference to LinkedHashMap
in the app, we were not able to apply method specializations in
HashMap because all specialized method was potentially polymorphic.
With the new optimization if the variable is only assigned to
HashMap instance, now we potentially identify that and specialize
the methods. Same is true for any optimizations that exploit the
fact that a method doesn't require polymorphic dispatch
(like inlining).

The implementation introduces refactors JNonNullType into JAnalysisType
whichn now can represent a non null subtype (or itself, tagged with
NOT_NULL) or a specific type (no subclass) that is also not null
(tagged with EXACT).
For example, if a JAnalysisType refers to HashMap and is tagged with
EXACT, the type cannot be any subtype but the HashMap itself.
The only places where the an EXACT JAnalysisType is a result of
newing a type (when a type is instantiated, tts precise type is known)
and then it is propagated mostly via TypeTightener.

Change-Id: Iad95138d336bbda66bbdbcc6d007b1b722d1c604
Review-Link: https://gwt-review.googlesource.com/#/c/10932/