Improves incremental compilation speed.
CompilationStateBuilder verifies contentId has not changed for
resources whose modification date has not changed, to warn the user.
This is very costly for incremental compile as most files are
not changed but we still need to read the content to verify that.
This was one of the obvious issues that showed up when I profiled
the compiler.
This patch puts contentId verification behind assert so that most
users can avoid the cost unless they hit an issue.

This patch also replaces costly isPackage check with a simpler
cache check while resolving types.

The patches cut hello world compilation time in my Mac by 30%
for incremental updates (~1.45s to ~0.95s).

Change-Id: I299768dbc7e881f2459d7bb84d33a0f5cc3f5799