Update patch set 3
Patch Set 3:

Having JsInteropExportsGenerator with two different implementation sound like a good direction to me as well.

For the testing, I'm not sure what is the best approach. We need to find something maintainable and low overhead. It would be great if it was easy to write a test for JsInteropExportsGenerator but probably that will not be easy.

Another option is to re-run the jsinterop tests with the closure flag by putting closure stub code in the host the html.

Patch-set: 3
Label: Verified=0