Add @JsOpaque feature.
The purpose is to prevent from passing unmatched variables
at JS side accidentally.

For example,
@JsType
interface MyJsInterface {
  void foo(A a); // A is a regular Java class or interface.
}
Here A doesn't have any accessible methods from javascript and
most of the time this is just a programming error. In some cases,
this might be intentional as the parameter will be only treated
as a opaque object (e.g. a Map). In that case the programmer
can annotate the attribute with @JsOpaque.

@JsOpaque applies the following rule:
Restrict the type of the following variables
[parameters of JsType method and exported method,
return of JsType method and exported method,
JsType field, exported field],
which are able to escape to JavaScript,
to be the following types:
[PrimitiveType, isOrExtendsJsType, isOrExtendsJsFunction,
JavaScriptObject, String, Object, Array of the above types,
and @JsOpaque Object].
Otherwise, it will be a compiler error (in current patch
it is a warning).

Change-Id: I6e3e3f08d36ce45bbf044314e3b5320005e34259