Make Closure output more idiomatic
The following changes result in an 11% improvement in Inbox when
compiling with closure.

In Closure mode, a synthetic constructor is generated to be passed
to the defineClass and then the generated prototype of the constructor
is explicitly assigned to real constructors.

All members are now assigned explicitly by setting on the prototype
of the synthesized constructor explicitly (e.g. Foo.prototoype.method)
instead of assigning via shared global (e.g. _.method).

This is because when the Closure compiler sees every method assigned
as _.method = expr;, it has no idea what '_' and how to associate
a given method with a given constructor. By changing this to
Foo.prototype.method = expr, the compiler can now associate
instantiations of Foo with liveness of methods assigned to it.

The GWT compiler now has responsibility for generating all
goog.provide statements and setting up empty functions on
goog.provided interface types.

The reason this is done is several fold. The closure compiler
will error if any goog.provide happens more than once, and it
will error if a goog.provide or a namespace is setup out of
order (enclosed type before enclosing type). Previously, a linker
woud generate all of the goog.provide statements as well as
setup stub functions for interfaces. However since linkers can't
(easily) insert code in the right order (like setting up a empty
interface stub function) in the topologically correct order with
those setup by the GWT compiler itself, pushing this reponsibility
to the compiler removes errors from bad sorting.

Also, double assigning any namespace causes deoptimization.
Previously the linker would pre-pend a bunch of empty stub
constructors which are written over by the main GWT module.

With this patch, now also Closure Mode correctly applies JsInterop
spec and reaches feature parity to non-Closure mode.

Finally, this is a step along the way to getting rid of the
linker, and having GWT output @JsDoc'ed class definitions.

Change-Id: I66aae7a745bb36059f1f0670d2820861527f2fc9