Update patch set 1
Patch Set 1:

I think your implementation should be same as getText/setText. I think it was like that before?

Your test is failing just because htmlunit is not aware of the placeholder so returns null. That particular part can be extracted as a different test and marked with DoNotRunWith(Platform.HtmlUnitBug)

Patch-set: 1
Label: Verified=0