Enhances TypeTightener with ability to lock into exact types.
Being able to lock to an exact type helps us to mark even more
methods as non-polymorphic which provides new optimization
opportunities.

Earlier, for example, if there was any reference to LinkedHashMap
in the app, we were not able to apply method specializations in
HashMap because all specialized method was potentially polymorphic.
With the new optimization if the variable is only assigned to
HashMap instance, now we potentially identify that and specialize
the methods. Same is true for any optimizations that exploit the
fact that a method doesn't require polymorphic dispatch
(like inlining).

The implementation introduces exactness to JAnalysisDecoratedType;
now a JAnalysisDecoratedType can be tagged as NULLABLE_NOT_EXACT,
NULLABLE_EXACT, NOT_NULLABLE_NOT_EXACT or NOT_NULLABLE_EXACT.

For example, if a JAnalysisType refers to HashMap and is tagged with
NULLABLE_EXACT, the object cannot be any subtype but the HashMap itself
or null.

The only places where the an *_EXACT JAnalysisType is a result of
newing a type (when a type is instantiated, its precise type is known)
and then it is propagated mostly via TypeTightener.

Change-Id: Ie1ee6731c5be249cca84f13c42f10027791a8cae