Provide minimum set of IntelliJ project files
After configuring GWT_ROOT and GWT_TOOLS path variables in IntelliJ the
"intellij" folder can be opened as IntelliJ project. The project has
preconfigured modules for gwt-user, gwt-dev and gwt-codeserver that
should compile without errors out of the box.

In addition IntelliJ is configured to:

* recognize GWT's ANT script so it can be executed inside the IDE
* recognize the project as GIT root
* use Google's Java code style
* automatically apply Apache 2.0 License header for new files
  (if license plugin is active)
* allow executing GWTTestCases right inside the IDE
* enable annotation processing for gwt-user with preconfigured
  requestfactory-apt.jar (must first be build manually using the
  "requestfactory" task of GWTs ANT script. Can be done inside the IDE)

The above does not require the IntelliJ GWT Plugin so it should also
work in the community edition of IntelliJ.

Change-Id: I7eb0a3f494ea689f878402f45f0bd5679fcb8cb2