Update patch set 1
Patch Set 1: Code-Review-1

I don't think this features is useful. Whenever I have used ScriptInjector the code to inject a script naturally ended up in some initialization method of my app/library that will only execute once anyways. This is especially true for FromUrl injections if a download has to be done and you have to wait for it. But thats subject to discussion so why do you need that feature in ScriptInjector?

I'll give this CL a "-1" for now not because of the above but because the implementation of this CL behaves bad (unless I miss something). If I create two ScriptInjectors with this feature enabled and each with its own callback then the first callback will wait until onload/onerror events have occurred but the second callback immediately says that the script has been loaded successfully which can be wrong. Thats because the registry gets filled with url + script element immediately without waiting for onload/onerror events.
At the end this can cause bugs in the app because it will execute code that depends on the injected script which might not has been downloaded and injected yet.

Patch-set: 1