Update patch set 2
Patch Set 2:

I think the fix is good, but I think in some cases you might actually want to generate a string message from SafeHtml inputs.  The alternate solution would be to make it work properly, but that does run the risk of people missing places they need to change to switch from String to SafeHtml, though I think compositing messages using the output of other messages is not frequent (and generally not a good idea for localization).

Thoughts on this solution vs. making the combination work properly?

Patch-set: 2