Update patch set 2
Patch Set 2:

> I think the fix is good, but I think in some cases you might
 > actually want to generate a string message from SafeHtml inputs. 
 > The alternate solution would be to make it work properly, but that
 > does run the risk of people missing places they need to change to
 > switch from String to SafeHtml, though I think compositing messages
 > using the output of other messages is not frequent (and generally
 > not a good idea for localization).
 > 
 > Thoughts on this solution vs. making the combination work properly?

Would you mind sharing the case when one would want to produce localized String which takes SafeHtml argument?
IMHO this is always a mistake: the argument is essentially HTML, but the output is a String (no HTML). Even though we could do some sort of conversion, this will mostly lead to confusion. Consider a case when developer does setText(MESSAGES.someMsg(makeLink(...)))
a) we include the HTML as verbatim text: user will see <a href='...'>...</a>
b) we take the text, stripping out the tags: user will see the text OK, but the link won't be clickable
I'd say that if you want to do either of the two conversions, you should code it explicitly (like safeHtml.asString()) instead of getting into trouble if you unintentionally forget about the SafeHtml during refactoring.
WDYT?

Patch-set: 2