Update patch set 1
Patch Set 1:

> I too thought this way, but I got a feedback regarding that-
 > https://github.com/gwtproject/gwt/pull/8#discussion-diff-27803052R274

Yes the feedback is correct but with your changes you still have the race condition I pointed out. When this features is enabled you must keep track of callbacks as long as the first download is still pending. Just remembering URLs is not enough to give correct answers to callbacks. Basically your solution is too simple.

When duplicates are disallowed the logic is more like:

   if (multipleDownloadsDisallowed) {
     if (isRequestPending(url)) {
       queueCallbackUntilRequestCompletes(url, callback);
     } else if (isScriptLoaded(registry, url)) {
       callback.onSuccess(null);
     } else {
       // when download is done: update registry, process callback queue for url.
       download(url, callback);
     }
   } else {
     download(url, callback);
   }

Patch-set: 1