Update patch set 1
Patch Set 1:

(7 comments)

This change causes a change in the behaviour of Tree, since title will now only be displayed on the actual text.

Does the tree work properly if the children have a title set other than empty string, or is this only an issue if children have an empty title?

Patch-set: 1