Update patch set 1
Patch Set 1:

> (7 comments)
 > 
 > This change causes a change in the behaviour of Tree, since title
 > will now only be displayed on the actual text.
 > 
 > Does the tree work properly if the children have a title set other
 > than empty string, or is this only an issue if children have an
 > empty title?

With the change, the tree works properly, in that 

 > (7 comments)
 > 
 > This change causes a change in the behaviour of Tree, since title
 > will now only be displayed on the actual text.
 > 
 > Does the tree work properly if the children have a title set other
 > than empty string, or is this only an issue if children have an
 > empty title?

Before the change, setting the root title set the title of the div containing the whole tree. This means that hovering on anything inside of the tree that did not have a title set on it (or a parent of it) would display the root title. 

With the change, you are correct, the title will only be displayed on the content of the treeitem. This means that hovering will only work for the actual content and not the container of the content, as it did before.

Is this acceptable behavior?

Patch-set: 1
Reviewer: Gerrit User 8530 <8530@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0