Update patch set 1
Patch Set 1:

(3 comments)

I changed the bridge methods to either just be an aliasing assignment (no long conversions needed) or to just delegate to the method and replace it's parameters with coerced parameters.

This is the wrong CL to remove this conversion feature. If we want to remove it, or implement it via @JsConvert it needs to be done in a later CL after changes are landed first to apps that are dependent on it.

Patch-set: 1
Label: Verified=0