Update patch set 1
Patch Set 1:

> Hi Daniel,
 > 
 > I didn't know about that. But for example take a look on
 > Collections.add(), the method is equal than the Oracle JRE only
 > variable's name are different.
 > GWT:
 > public static <T> boolean addAll(Collection<? super T> c, T... a) {
 > boolean result = false;
 > for (T e : a) {
 > result |= c.add(e);
 > }
 > return result;
 > }
 > 
 > Oracle JDK:
 > public static <T> boolean addAll(Collection<? super T> c, T...
 > elements) {
 > boolean result = false;
 > for (T element : elements)
 > result |= c.add(element);
 > return result;
 > }
 > 
 > So knowing that. Is it acceptable apply the same to rotate method?

No it is not acceptable.

In the case of addAll there are not that many ways to write it so it might be a coincidence that its control flow looks the same with different variable names. 

For Collections.rotate however there are many ways in which you can write it.

It is never OK to copy a JRE method.

GWT implementations should be written from scratch.

Patch-set: 1
Reviewer: Gerrit User 5105 <5105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0