Update patch set 1
Patch Set 1:

> Change has been successfully rebased as ccdc7254409a8c4705a084618e5660c2efc9ae1d
 > by Goktug Gokdogan

What is the main reason to remove this feature which actually was already working in current implementation. 
We at Vaadin have a lot of code in our new products using chaining  and properties without get/set which is broken right now, so we will have to modify, or trust in a frozen snapshot previous to this which is not optimal.

Patch-set: 1
Label: Verified=0