Update patch set 3
Patch Set 3:

> > Change has been successfully rebased as ccdc7254409a8c4705a084618e5660c2efc9ae1d
 > > by Goktug Gokdogan
 > 
 > What is the main reason to remove this feature which actually was
 > already working in current implementation.
 > We at Vaadin have a lot of code in our new products using chaining 
 > and properties without get/set which is broken right now, so we
 > will have to modify, or trust in a frozen snapshot previous to this
 > which is not optimal.

Actually I described it in the commit message but let me give more explanation.

JsInterop was an experimental feature and we are preparing it towards 1.0 release. That is why we hide it behind experimental flag. We are trimming of parts that are not finalized or if we are not sure to support it on the long run (You can add stuff later but it is difficult to remove stuff later).

The issue specific to supporting JsProperty is, things gets really hairy when you support many different getter/setter styles. For example, unlike our initial implementation, @AutoValue supports only single style per class. That makes a lot of sense to me. However, we also have inheritance unlike @AutoValue, so how will the inheritance hierarchy be effected by this? What happens if I implement two different interfaces that follows different semantics?

Also in AutoValue, if there is single non java-bean style method, then the processor considers everything as non-java bean style. This is very error-prone as if somebody adds a new property without paying attention to the rest it will just switch the meaning of every JsProperty in the class.

W.r.t fluent chaining; if you think about the concrete classes, then the things gets hairy again as they now need to return values for their setter methods. So if we ever had such future, I think it should be thought out as a separate helper like @AutoValue.

Also note that; javascript doesn't have such feature either. If the javascript API already provides a fluent API then you have a fluent API in GWT side as well. By wrapping it with auto-conversion to fluent style, we are basically trying to fix a broken javascript API.

If somebody really needs this, they can simply introduce an annotation processor, e.g. @JsTypeBuilder, outside of the SDK (similar to google/auto projects).

Patch-set: 3