Changes the export naming of constructors.
Earlier a constructor not qualified with a name would have
export namespace as the class' namespace and export name
as the class' name.
With this change the export name is now empty and the
namespace is the fully qualified name of the class.

Also now if the constructor provides simple name, the fully
qualified now resides under the class' qualified name.
(e.g. for class a.b.C, constructor named 'd' will have the
fully qualified name a.b.C.d)

This is a step towards handling of constructors as factory
methods when there are multiple constructors exist.

Change-Id: I4d2b37eff15b718c1aa7b7b4797c5729c5f66f93