Changes the export naming of constructors.
Earlier a constructor fully qualified as "a.b.C" would have
export namespace 'a.b' and export name 'a.b'. With this change
the export namespace is 'a.b.C' and the name is empty.

This patch doesn't introduce any user visible change but changes
how we keep export data internally. It will help further changes
towards handling of constructors as factory methods when there
are multiple constructor exist.

Change-Id: I4d2b37eff15b718c1aa7b7b4797c5729c5f66f93