Update patch set 3
Patch Set 3:

Is there a text class for a class with a single exported constructor without a name provided, and if not, could you add it? e.g.


@JsExport
class MyClassWithDefaultCtor {
}

or

class MyClassWithOneCtor {
   @JsExport
   public MyClassWithOneCtor() {
   }
}

I know we discussed eventually moving to an all factory method approach, but I want to ensure current code doesn't break until we refactor.

Patch-set: 3