Converts multi ctors to factory methods for JsExport.
Exporting multiple constructors for the same object in different
namespace is problematic due to js instanceof checks and placing
static methods. Also in the long run, we are already moving towards
generating single constructor per class while generating user-defined
constructors as factory methods so this is making JsInterop inline
with that.

With this patch, multiple constructors will be exported as factory
methods instead of real constructors so one needs to supply a name for
it.
If there is only one constructor, it can be still exported without
supplying a name and will be exported as a regular constructor.

Change-Id: I4008afe6d86261c6de2c7bda3aa1dfad992123a4