Fix FragmentExtractor for -XclosureFormattedOutput mode
In closure output mode, the chain assignments of ctors are
always retained in the initial fragment, while their constructors
may be moved to a later fragment, causing undefined references.

This change properly moves the chain setup to the fragment where
they are needed.

Also adds clinit() call to AsyncFragmentLoader.onLoad() which is
assumed always to be called, but in fact, depending on the
LoadingStrategy, the onLoad() method may be invoked before the
clinit() is ever triggered somewhere else. (In the case where
an 'already loaded' loading strategy is used, where a server-side
linking process concatenates the output onto the initial download)

Change-Id: Ic2f7f90a287bf11fe9e04dec2fdd6f132fda98b1