Update patch set 2
Patch Set 2:

Sent https://gwt-review.googlesource.com/#/c/12692/ that reverts the ones in the SDK as it has been blocking our compiler code size tests.

By the way, I forgot who were going to send the meeting notes from the last week. As the time passes, we will forget about the details so it would be great if you guys can send it before that =)

Patch-set: 2