Update patch set 1
Patch Set 1:

I've take the decision to not enable GSS by default in uibinder test because we have tests for GSS and uibinder, so that we test both the integration of GSS and the old syntax with uibinder.

If we enable GSS for uibinder tests we have to write tests for testing that the old css syntax is still working.

Patch-set: 1