Update patch set 1
Patch Set 1:

I don't have access to the code so I'm thinking aloud: with this solution do we have still tests for testing uibinder with gss disabled? It was the main why I didn't fix the test with this way.

Another alternative is to simply force gss to be disabled for the uibinder tests.

Patch-set: 1