Update patch set 1
Patch Set 1:

Commiting this means that we are only running tests for UiBinder and gss.

In the last meeting about GSS we said its not default for 2.8, but we are not going to bother fixing any upcoming issues for old CSSResource. So I think we are good not running these tests in css mode anymore.

Patch-set: 1