Update patch set 1
Patch Set 1:

Ideally we need tests both way (both with gss disabled and gss enabled). There are both beneficial to test but also probably an overkill. Google currently running everything with GSS enabled so we got that covered. If we keep this running with old CSS we will get that covered as well; otherwise we will be blind to any breakages to old css as we had a few weeks back.

Patch-set: 1