Update patch set 1
Patch Set 1:

I'll add more tests. In the meantime, to answer the general arguments to the approach about being too conservative: I disagree.

The logic is quite simple. If a method call to JsType is reachable, it needs to be rescued. It doesn't matter if the object is ended up there because of cast or through native methods or through an exported method. Instead of tracking all these entry points we simply rescue the type when we see the call.

If you can show me a scenario where a type is 'unnecessary' rescued with this simple logic, I can look for ways to improve that but otherwise this seems quite precise to me.

Patch-set: 1