Update patch set 1
Patch Set 1:

That's my point we have already tests in the sdk that cover the usage of uibinder and gss. So we have to disable gss in these test to cover case where gss is disabled. By setting :
  
    <set-configuration-property name="CssResource.enableGss" value="false" />

in the module file we doesnt depend of the default configuration anymore

Patch-set: 1