Update patch set 1
Patch Set 1:

> We recently just found a bug in our uibinder integration just
 > because I switched these test to run on gss, so our current gss
 > tests are not covering the same amount.
 > 
 > I would prefer these tests running gss very much.
 > 
 > @Goktug if we decide to run them in css and I import the patch
 > google won't be running them in gss either.

If we explicitly set them to run with css in gwt.xml, google3 version will be running in css as well.

Ok, as a middle ground, what about not importing the change to google3 so that our internal tests will run with gss and jenkins will keep running with css. Does it sound good to everybody?

Patch-set: 1