Fixes JsFunction to be instance of Function.
This makes the generated functions from JsFunction
to behave more like a real function.

In order to achieve that we replaced the parent of the class
to be the javascript Function.prototype. This requires us to
not support extending other classes from JsFunction classes.

Also we prevented other inheritance scenarios like implementing
multiple interfaces from the implementation so that in the future
we can replace the generated code to be a real function that is
not __proto__ treated by putting the instance methods to the
generated function instead.

Change-Id: I4dcb519cf96887072d5a1771fe891574d7d6b101