Update patch set 2
Patch Set 2:

I think this code shouldn't be behind a flag. When interfaces are empty no code will be added to the output anyway and now with Java 8 are not always empty. 

I think the simplest way is to consider interfaces as instantiated when a type implementing them is instantiated, at the end of RTA.markTypeIdInstanted.

Patch-set: 2
Reviewer: Gerrit User 5105 <5105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0