Fixes incorrect obfuscation of names for JsInterop.
As old browsers doesn't support keywords as property names
GWT may mangle the name even if it is marked as unobfuscatable.
The only browser that doesn't support this is IE8[1] and
IE8 is not supported by JsInterop.

Our native methods already cannot contain such keywords as
our parser rejects it. The rest of the non-obfuscated names
are coming from programmatic name generation and we investigated
code to ensure that all names are either already mangled or
marked as obfuscated.

So the patch simply removes the mangling.

Alternative could be using braced property access (obj["..."])
whenever the property/function name is a keyword.
That would at least let IE8 successfully parse the script though
that would be a much larger change.

[1] http://kangax.github.io/compat-table/es5/#Reserved_words_as_property_names

Change-Id: If55965b09f2f7892cabf88830c1fa91f59ef4f22
Review-Link: https://gwt-review.googlesource.com/#/c/11790/